Init sql data ok, Hibernate working, Trainees Test ok

+++ src/main/java/com/epam/dalvaradoc/mod2_spring_core_task/utils/UserUtils.java
@@ -13,7 +13,7 @@ public class UserUtils {
   public static String getSaltString() {
     StringBuilder salt = new StringBuilder();
     while (salt.length() < 10) { // length of the random string.
-      int index = (rnd.nextInt() * SALTCHARS.length());
+      int index = (int) (rnd.nextFloat() * SALTCHARS.length());
       salt.append(SALTCHARS.charAt(index));
     }
     return salt.toString();
